#8958 - Resolve Versioned dev/build performance with large datasets
During a dev/build Silverstripe calls an extension augmentDatabase()
on Versioned DataObjects. During this function a cleanup query is
made which was supposed to cleanup instances where the uniqueness
constraint can't be used.

This results in terrible performance as it's executed on every build.

This fix moves that functionality to a task, so it can be cleaned up
and added to a cron should it be deemed necessary. It's not currently
known if the bug still manifests.

Also removed some commented code.